Make the command line a client of the service
Every command now calls org.nspawn, the way machinectl and systemctl
are clients of machined and systemd, so one code path does the work.
The registry and CA certificate the command line was given travel as
options of each call, -e VAR is expanded from the caller's environment
before the call, jobs are followed line by line through JobOutput,
and exec, shell and logs attach the terminal to the descriptors the
service hands over. Only the service, its installation and the unit
hooks run the library in the command line's own process.

The service gains Shell, per-call registry options, a kind on
JobOutput and mkosi's output as job lines. The suite installs the
service first; a start issued while a failed program is still alive
is correctly refused as already running, and the step allows for it.

+++ src/daemon/values.rs
@@ -101,7 +101,7 @@ pub fn machine(m: &MachineSummary) -> Dict {
     dict.insert("name".to_string(), v(m.name.as_str()));
     dict.insert("state".to_string(), v(m.state.as_str()));
     dict.insert("started".to_string(), v(m.started.unwrap_or(0)));
-    dict.insert("leader".to_string(), v(m.leader.unwrap_or(0)));
+    dict.insert("leader".to_string(), v(u64::from(m.leader.unwrap_or(0))));
     dict.insert("os".to_string(), opt_string(m.os.as_deref()));
     dict.insert(
         "machine_path".to_string(),
